+++ amazonorders/cli.py
@@ -15,10 +15,9 @@ from click.core import Context
 
 from amazonorders import __version__, util
 from amazonorders.conf import AmazonOrdersConfig
-from amazonorders.entity.order import Order
-from amazonorders.entity.transaction import Transaction
 from amazonorders.exception import AmazonOrdersError, AmazonOrdersAuthError, AmazonOrdersAuthRedirectError
 from amazonorders.orders import AmazonOrders
+from amazonorders.output import OutputFormatter
 from amazonorders.session import AmazonSession, IODefault
 from amazonorders.transactions import AmazonTransactions
 
@@ -128,6 +127,8 @@ def amazon_orders_cli(ctx: Context,
                    "This will execute an additional request per Order.")
 @click.option("--order-filter", "order_filter", default=None,
               help="The order type filter to use.")
+@click.option("-o", "--output", type=click.Choice(OutputFormatter.OUTPUT_FORMATS), default="text",
+              help="The output format. Defaults to text.")
 def history(ctx: Context,
             **kwargs: Any) -> None:
     """
@@ -145,6 +146,7 @@ def history(ctx: Context,
         single_page = kwargs["single_page"]
         full_details = kwargs["full_details"]
         order_filter = kwargs["order_filter"]
+        output = kwargs["output"]
 
         exclusive_flags = [year, last_3_months, last_30_days]
         if sum(1 for item in exclusive_flags if item) > 1:
@@ -170,28 +172,27 @@ Order History for {filter_description}{optional_start_index}{optional_full_detai
 -----------------------------------------------------------------------\n"""
                    .format(filter_description=filter_description,
                            optional_start_index=optional_start_index,
-                           optional_full_details=optional_full_details))
-        click.echo("Info: Fetching Order history, this might take a minute ...")
+                           optional_full_details=optional_full_details), err=True)
+        click.echo("Info: Fetching Order history, this might take a minute ...", err=True)
 
         config = ctx.obj["conf"]
         amazon_orders = AmazonOrders(amazon_session,
                                      config=config)
 
         start_time = time.time()
-        total = 0
-        for o in amazon_orders.get_order_history(year=year,
+        orders = amazon_orders.get_order_history(year=year,
                                                  start_index=start_index,
                                                  full_details=full_details,
                                                  keep_paging=not single_page,
                                                  time_filter=time_filter,
-                                                 order_filter=order_filter):
-            click.echo(f"{_order_output(o, config)}\n")
-            total += 1
+                                                 order_filter=order_filter)
         end_time = time.time()
 
+        click.echo(config.output_cls(config).format(orders, output))
+
         click.echo(
-            "... {total} Orders parsed in {time} seconds.\n".format(total=total,
-                                                                    time=int(end_time - start_time)))
+            "... {total} Orders parsed in {time} seconds.\n".format(total=len(orders),
+                                                                    time=int(end_time - start_time)), err=True)
     except AmazonOrdersAuthRedirectError:
         _prompt_to_reauth_flow()
     except AmazonOrdersError as e:
@@ -202,8 +203,11 @@ Order History for {filter_description}{optional_start_index}{optional_full_detai
 @amazon_orders_cli.command()
 @click.pass_context
 @click.argument("order_id")
+@click.option("-o", "--output", type=click.Choice(OutputFormatter.OUTPUT_FORMATS), default="text",
+              help="The output format. Defaults to text.")
 def order(ctx: Context,
-          order_id: str) -> None:
+          order_id: str,
+          **kwargs: Any) -> None:
     """
     Get the full details for a given Amazon Order ID.
     """
@@ -216,9 +220,11 @@ def order(ctx: Context,
         amazon_orders = AmazonOrders(amazon_session,
                                      config=config)
 
+        output = kwargs["output"]
+
         o = amazon_orders.get_order(order_id)
 
-        click.echo(f"{_order_output(o, config)}\n")
+        click.echo(config.output_cls(config).format([o], output))
     except AmazonOrdersAuthRedirectError:
         _prompt_to_reauth_flow()
     except AmazonOrdersError as e:
@@ -256,8 +262,11 @@ def invoice(ctx: Context,
 @amazon_orders_cli.command("order-transactions")
 @click.pass_context
 @click.argument("order_id")
+@click.option("-o", "--output", type=click.Choice(OutputFormatter.OUTPUT_FORMATS), default="text",
+              help="The output format. Defaults to text.")
 def order_transactions(ctx: Context,
-                       order_id: str) -> None:
+                       order_id: str,
+                       **kwargs: Any) -> None:
     """
     Get the Transactions for a given Amazon Order ID.
     """
@@ -270,16 +279,17 @@ def order_transactions(ctx: Context,
         amazon_transactions = AmazonTransactions(amazon_session,
                                                  config=config)
 
+        output = kwargs["output"]
+
         start_time = time.time()
-        total = 0
-        for t in amazon_transactions.get_transactions(order_id=order_id):
-            click.echo(f"{_transaction_output(t, config)}\n")
-            total += 1
+        order_transaction_list = amazon_transactions.get_transactions(order_id=order_id)
         end_time = time.time()
 
+        click.echo(config.output_cls(config).format(order_transaction_list, output))
+
         click.echo(
-            "... {total} Transactions parsed in {time} seconds.\n".format(total=total,
-                                                                          time=int(end_time - start_time)))
+            "... {total} Transactions parsed in {time} seconds.\n".format(total=len(order_transaction_list),
+                                                                          time=int(end_time - start_time)), err=True)
     except AmazonOrdersAuthRedirectError:
         _prompt_to_reauth_flow()
     except AmazonOrdersError as e:
@@ -291,6 +301,8 @@ def order_transactions(ctx: Context,
 @click.pass_context
 @click.option("--days", default=365,
               help="The number of days of Transactions to get.")
+@click.option("-o", "--output", type=click.Choice(OutputFormatter.OUTPUT_FORMATS), default="text",
+              help="The output format. Defaults to text.")
 def transactions(ctx: Context, **kwargs: Any):
     """
     Get Amazon Transaction history for a given number of days.
@@ -301,28 +313,28 @@ def transactions(ctx: Context, **kwargs: Any):
         _authenticate(amazon_session)
 
         days = kwargs["days"]
+        output = kwargs["output"]
 
         click.echo(
             """-----------------------------------------------------------------------
 Transaction History for {days} days
------------------------------------------------------------------------\n""".format(days=days)
+-----------------------------------------------------------------------\n""".format(days=days), err=True
         )
-        click.echo("Info: Fetching Transaction history, this might take a minute ...")
+        click.echo("Info: Fetching Transaction history, this might take a minute ...", err=True)
 
         config = ctx.obj["conf"]
         amazon_transactions = AmazonTransactions(amazon_session,
                                                  config=config)
 
         start_time = time.time()
-        total = 0
-        for t in amazon_transactions.get_transactions(days=days):
-            click.echo(f"{_transaction_output(t, config)}\n")
-            total += 1
+        transaction_list = amazon_transactions.get_transactions(days=days)
         end_time = time.time()
 
+        click.echo(config.output_cls(config).format(transaction_list, output))
+
         click.echo(
-            "... {total} Transactions parsed in {time} seconds.\n".format(total=total,
-                                                                          time=int(end_time - start_time)))
+            "... {total} Transactions parsed in {time} seconds.\n".format(total=len(transaction_list),
+                                                                          time=int(end_time - start_time)), err=True)
     except AmazonOrdersAuthRedirectError:
         _prompt_to_reauth_flow()
     except AmazonOrdersError as e:
@@ -400,7 +412,7 @@ def version(ctx: Context) -> None:
 
 
 def _print_banner() -> None:
-    click.echo(banner.format(version=__version__))
+    click.echo(banner.format(version=__version__), err=True)
 
 
 def _authenticate(amazon_session: AmazonSession,
@@ -437,57 +449,6 @@ def _authenticate(amazon_session: AmazonSession,
 def _prompt_to_reauth_flow() -> None:
     click.echo("... Amazon redirected to login, which likely means the persisted session is stale. It was logged "
                "out, so try running the command again.\n")
-
-
-def _order_output(o: Order,
-                  config: AmazonOrdersConfig) -> str:
-    order_str = """-----------------------------------------------------------------------
-Order #{order_number}
------------------------------------------------------------------------""".format(order_number=o.order_number)
-
-    order_str += f"\n  Shipments: {o.shipments}"
-    order_str += f"\n  Order Details Link: {o.order_details_link}"
-    if o.grand_total:
-        order_str += f"\n  Grand Total: {config.constants.format_currency(o.grand_total)}"
-    order_str += f"\n  Order Placed Date: {o.order_placed_date}"
-    if o.recipient:
-        order_str += f"\n  {o.recipient}"
-    else:
-        order_str += "\n  Recipient: None"
-
-    if o.payment_method:
-        order_str += f"\n  Payment Method: {o.payment_method}"
-    if o.payment_method_last_4:
-        order_str += f"\n  Payment Method Last 4: {o.payment_method_last_4}"
-    if o.subtotal:
-        order_str += f"\n  Subtotal: {config.constants.format_currency(o.subtotal)}"
-    if o.shipping_total:
-        order_str += f"\n  Shipping Total: {config.constants.format_currency(o.shipping_total)}"
-    if o.free_shipping:
-        order_str += f"\n  Free Shipping: {config.constants.format_currency(o.free_shipping)}"
-    if o.subscription_discount:
-        order_str += f"\n  Subscription Discount: {config.constants.format_currency(o.subscription_discount)}"
-    if o.total_before_tax:
-        order_str += f"\n  Total Before Tax: {config.constants.format_currency(o.total_before_tax)}"
-    if o.estimated_tax:
-        order_str += f"\n  Estimated Tax: {config.constants.format_currency(o.estimated_tax)}"
-    if o.refund_total:
-        order_str += f"\n  Refund Total: {config.constants.format_currency(o.refund_total)}"
-
-    order_str += "\n-----------------------------------------------------------------------"
-
-    return order_str
-
-
-def _transaction_output(t: Transaction,
-                        config: AmazonOrdersConfig) -> str:
-    transaction_str = f"Transaction: {t.completed_date}"
-    transaction_str += f"\n  Order #{t.order_number}"
-    if t.grand_total:
-        transaction_str += f"\n  Grand Total: {config.constants.format_currency(t.grand_total)}"
-    transaction_str += f"\n  Order Details Link: {t.order_details_link}"
-
-    return transaction_str
 
 
 if __name__ == "__main__":